Do not modify the "default" profile
It turns out that setting the session token on the default profile is a bad idea.
It works the first time it's updated, when there's no existing session token.
After that we get a "The security token included in the request is invalid." error every time.

Now the profile parameter is no longer optional and it must be set to
something other than "default".